%c + изменены int на long long в структурах

+++ src/s21_sprintf.h
@@ -9,33 +9,33 @@
 // {-} {+} {' '} {#} {0}
 // количество вхождений флага
 typedef struct _flags {
-    int minus;
-    int plus;
-    int space;
-    int hash;
-    int zero;
+    long long minus;
+    long long plus;
+    long long space;
+    long long hash;
+    long long zero;
 } fl;
 
 // {0..9} {*} 
 // количество вхождений флага ширины
 typedef struct _width {
-    int number;
-    int starchar;
+    long long number;
+    long long starchar;
 } w;
 
 // {0..9} {*} 
 // количество вхождений флага точности
 typedef struct _precision {
-    int number;
-    int starchar;
+    long long number;
+    long long starchar;
 } pr;
 
 // {h} {l} {L}
 // количество вхождений флага длины.
 typedef struct _length {
-    int h;
-    int l;
-    int L;
+    long long h;
+    long long l;
+    long long L;
 } len;
 
 typedef struct _format{
@@ -50,4 +50,4 @@ typedef struct _format{
 int s21_sprintf(char *str, const char *format, ...);
 
 //  позиция указателя до конца подстроки замещаемой части формата
-int endfmt(const char *str, int start);
+int endfmt(const char *str, long long start);
